finished 1-list_len

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -1,9 +1,15 @@
 #include <stdio.h>
 #include "lists.h"
-
+/**
+ * list_len - return length of list
+ * @h: pointer to list_t(struct list_s)
+ * Return: size_t
+ *
+ * Description: traverse list and return size
+ */
 size_t list_len(const list_t *h)
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (h == NULL)
 		return (0);
